LayoutTests/plugins/geturlnotify-during-document-teardown.html

<html>
    <head>
        <script>
        if (!window.layoutTestController) {
            alert('this test can only be run by DumpRenderTree')
        } else {
            layoutTestController.dumpAsText();
            layoutTestController.waitUntilDone();
            layoutTestController.setCallCloseOnWebViews(false);
            layoutTestController.setCanOpenWindows();
            layoutTestController.setCloseRemainingWindowsWhenComplete(false);
            window.open("resources/geturlnotify-on-destroy.html");
        }
        </script>
    </head>
    <body>
        This tests that performing a load during document teardown does not
        cause a crash.
        <a href="https://bugs.webkit.org/show_bug.cgi?id=38797">Bug #38797</a>
    </body>
</html>
