LayoutTests/plugins/geturlnotify-during-document-teardown-expected.txt

CONSOLE MESSAGE: line 0: PLUGIN: NPP_Destroy
This tests that performing a load during document teardown does not cause a crash. Bug #38797
